+++ packages/@expo/cli/src/start/server/metro/__tests__/createExpoMetroResolver.test.ts
@@ -1,7 +1,7 @@
 import { getBareExtensions } from '@expo/config/paths';
 import assert from 'assert';
 import fs from 'fs';
-import type { SourceFileResolution } from 'metro-resolver/src/types';
+import type { SourceFileResolution } from 'metro-resolver/private/types';
 import path from 'path';
 
 import { createFastResolver, FailedToResolvePathError } from '../createExpoMetroResolver';
