+++ packages/@expo/cli/src/start/server/metro/__tests__/metroErrors.test.ts
@@ -1,5 +1,5 @@
-import FailedToResolveNameError from 'metro-resolver/src/errors/FailedToResolveNameError';
-import FailedToResolvePathError from 'metro-resolver/src/errors/FailedToResolvePathError';
+import FailedToResolveNameError from 'metro-resolver/private/errors/FailedToResolveNameError';
+import FailedToResolvePathError from 'metro-resolver/private/errors/FailedToResolvePathError';
 
 import { isFailedToResolveNameError, isFailedToResolvePathError } from '../metroErrors';
