+++ packages/@expo/cli/src/start/server/metro/getCssModulesFromBundler.ts
@@ -1,10 +1,10 @@
 import { MetroConfig } from '@expo/metro-config';
 import crypto from 'crypto';
 import type { Module } from 'metro';
-import { getJsOutput, isJsModule } from 'metro/src/DeltaBundler/Serializers/helpers/js.js';
-import type { ReadOnlyDependencies } from 'metro/src/DeltaBundler/types';
-import type IncrementalBundler from 'metro/src/IncrementalBundler';
-import splitBundleOptions from 'metro/src/lib/splitBundleOptions';
+import { getJsOutput, isJsModule } from 'metro/private/DeltaBundler/Serializers/helpers/js.js';
+import type { ReadOnlyDependencies } from 'metro/private/DeltaBundler/types';
+import type IncrementalBundler from 'metro/private/IncrementalBundler';
+import splitBundleOptions from 'metro/private/lib/splitBundleOptions';
 import path from 'path';
 
 type Options = {
