+++ packages/@expo/cli/src/start/server/metro/metroPrivateServer.ts
@@ -4,19 +4,19 @@ import { ConfigT } from 'metro-config';
 import assert from 'node:assert';
 
 export type MetroPrivateServer = import('metro').Server & {
-  _bundler: import('metro/src/IncrementalBundler').default;
+  _bundler: import('metro/private/IncrementalBundler').default;
   _config: ConfigT;
   _createModuleId: (path: string, context?: { environment?: string; platform: string }) => number;
   _isEnded: boolean;
   _nextBundleBuildNumber: number;
   _platforms: Set<string>;
-  _reporter: import('metro/src/lib/reporting').Reporter;
+  _reporter: import('metro/private/lib/reporting').Reporter;
   _serverOptions: import('metro').ServerOptions | void;
 
   getNewBuildNumber(): number;
   _getSortedModules(
-    graph: import('metro/src/IncrementalBundler').OutputGraph
-  ): import('metro/src/DeltaBundler/types').Module[];
+    graph: import('metro/private/IncrementalBundler').OutputGraph
+  ): import('metro/private/DeltaBundler/types').Module[];
 
   _resolveRelativePath(
     filePath: string,
@@ -26,12 +26,12 @@ export type MetroPrivateServer = import('metro').Server & {
       transformOptions,
     }: {
       relativeTo: 'project' | 'server';
-      resolverOptions: import('metro/src/shared/types').ResolverInputOptions;
+      resolverOptions: import('metro/private/shared/types').ResolverInputOptions;
       transformOptions: TransformInputOptions;
     }
   ): Promise<string>;
 
-  _shouldAddModuleToIgnoreList(module: import('metro/src/DeltaBundler/types').Module<any>): boolean;
+  _shouldAddModuleToIgnoreList(module: import('metro/private/DeltaBundler/types').Module<any>): boolean;
 };
 
 export function assertMetroPrivateServer(metro: Metro.Server): asserts metro is MetroPrivateServer {
