+++ packages/@expo/cli/src/start/server/metro/runServer-fork.ts
@@ -7,8 +7,8 @@ import assert from 'assert';
 import http from 'http';
 import https from 'https';
 import Metro, { RunServerOptions, Server } from 'metro';
-import MetroHmrServer from 'metro/src/HmrServer';
-import createWebsocketServer from 'metro/src/lib/createWebsocketServer';
+import MetroHmrServer from 'metro/private/HmrServer';
+import createWebsocketServer from 'metro/private/lib/createWebsocketServer';
 import { ConfigT } from 'metro-config';
 import { parse } from 'url';
 import type { WebSocketServer } from 'ws';
